requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
phonenumbers==8.13.55